Updated migrations
Changed datatype from integer to small int in reviewers & reviews for
corresponding rank & category. Rmoved timestamp from likes table

+++ database/migrations/2017_04_09_183826_create_reviewers_table.php
@@ -31,7 +31,7 @@ class CreateReviewersTable extends Migration
             $table->text('about')->nullable();
 
 
-            $table->integer('rank_id');
+            $table->smallInteger('rank_id');
             $table->timestamps();
 
             $table->rememberToken();
