+++ database/migrations/2017_05_09_212934_create_reviews_table.php
@@ -19,7 +19,7 @@ class CreateReviewsTable extends Migration
             $table->text('caption');
             $table->longText('body');
 
-            $table->integer('category_id');
+            $table->smallInteger('category_id');
             $table->integer('reviewer_id');
             $table->timestamps();
         });
